requirements-pip.txt
django
djangorestframework
django-jsonfield
django-jsonfield-backport
django-unixtimestampfield
django-cors-headers
hexdump
Crypto
pycrypto

